Fold Python's imports with VIM / SimpylFold
A new feature to fold/group the Python imports (like the IDE Pycharm):
    from module import name
    import module

Folding the imports is useful because Python programmers spend more time
in the classes/methods rather than imports. It is always interesting to
fold the imports and unfold them only when we need to import a new
module in our Python source code.

For those who don't want to fold their Python imports, they can use this
option:
    let g:SimpylFold_fold_import = 0