Support for Function Names and Docstrings
Added two variables which allow users not to fold function names (à la foldmethod=indent) and also allow the docstrings not to be folded. There are a couple of considerations I should put into the README, which are both new variables need default values, and also if you want to use unfold_docstring, unfold_function name should also be set to enable.